style(atlas): manual clang-tidy cleanup — remaining 95 warnings
Follow-up to 75ac751 which applied the auto-fixable clang-tidy checks
to the three atlas files. This commit handles the ~95 warnings that
required manual judgement:

bugprone-implicit-widening-of-multiplication-result (~30)
  - Cast first operand to (size_t) where the result is used as a
    pointer offset: (size_t)n * sizeof(T), (size_t)i * 2, etc.
  - Convert loop counters to size_t where they're only used as
    indices (for (size_t v = 0; v < n; v++)).
  - Change nfp_cache indexing to (size_t)cache_set_idx * VPACK_NFP_CACHE_WAYS.

cert-err33-c (~30)
  - (void) cast on tinycthread mtx_lock/mtx_unlock/cnd_wait/
    cnd_broadcast/cnd_signal/cnd_init/mtx_init/thrd_join return
    values. On our platforms (Windows/POSIX) the only failure modes
    are OOM or EINVAL on a destroyed primitive, neither of which is
    recoverable mid-batch. Comment added on the first occurrence.

readability-function-cognitive-complexity (13)
  - NOLINTNEXTLINE with a one-line explanation on:
      blit_sprite (fast-path + generic path)
      extrude_edges (four symmetric edge-copy loops)
      pipeline_register (multi-type entry emission)
      nt_builder_end_atlas (pipeline orchestrator)
      hull_simplify_perp (O(n²) greedy removal)
      polygon_max_outside_pixel_distance (pixel-grid scan)
      binary_build_convex_polygon (iterative convex refinement)
      rdp_simplify (closed-loop RDP with double-pivot)
      polygon_inflate (Clipper2 bridge plumbing)
      vpack_intersect_axis_i (symmetric x/y-axis branches)
      vpack_try_mark_cand_seen (hot-path dedup bitset)
      vpack_scan_candidate_range (NFP collision + score reduction)
      test_atlas_round_trip_basic (unfolded round-trip assertions)
  All represent inherent algorithm complexity, not tangled control
  flow — splitting would just shuffle locals across functions.

misc-no-recursion (1)
  - NOLINT on rdp_recurse: Ramer-Douglas-Peucker is defined
    recursively; an iterative version would need an explicit stack
    and be less clear.

bugprone-multi-level-implicit-pointer-conversion (4)
  - Explicit (void *) cast on free() calls where the pointer type
    is ** (Point2D **, char **, thrd_t *).

clang-analyzer-* (9)
  - codegen.c: switch collision-check sorted array to calloc so the
    analyzer sees every field initialised before qsort. Add explicit
    free(ce) in the fopen-failure path of nt_builder_generate_header
    (real memory leak). Guard write_sorted_defines and
    write_register_labels against count == 0 (real malloc(0)). Add
    NOLINT on one false-positive "uninitialized 5th argument" to
    NT_LOG_ERROR macro that analyzer loses through qsort.
  - atlas.c: NT_BUILD_ASSERT(sprite_count > 0) before malloc in
    pipeline_serialize so the analyzer can see sprite_count != 0.
    NOLINT on one "memory leak at return false" false positive in
    atlas_cache_read (no allocation is alive at that return).
    NOLINT on one "garbage operand" false positive in
    pack_stats_measure_payload (trim_w/trim_h are populated by
    the caller through pipeline_tile_pack, analyzer can't trace).
  - vpack.c: NOLINT on two "garbage operand of '-'/'+'" false
    positives where out-parameters from vpack_calc_aabb are used in
    the next expression — analyzer misses the out-parameter write.
    Convert one dead store of nfp_ok to (void)nfp_ok with a comment.

readability-uppercase-literal-suffix (1), readability-redundant-*
(several), google-readability-casting (several)
  - Mechanical cleanup: 1u → 1U, drop redundant (int32_t)
    casts, remove duplicate forward declaration of
    nt_builder_glob_iterate from nt_builder_internal.h (already in
    nt_builder.h), replace nested ternary in build_packs.c with
    if/else, switch atoi → strtoul (cert-err34-c), add NOLINT on
    getenv "thread-unsafe" warning in the single-threaded CLI tool.

Verified:
  - bash scripts/tidy.sh → all project checks passed
  - clang-format --dry-run --Werror → clean
  - test_builder native-debug → 105/105 PASS in ~1s
  - nt_builder + build_atlas_packs build cleanly

+++ tests/unit/test_builder.c
@@ -3985,22 +3985,22 @@ void test_atlas_real_pipeline_preserves_hole(void) {
              * this transparent pixel is enclosed — i.e. inside a hole. */
             bool blocked_n = false, blocked_s = false, blocked_w = false, blocked_e = false;
             for (uint32_t k = 1; y >= k && !blocked_n; k++) {
-                if (page_pixels[((size_t)(y - k) * page_w + x) * 4 + 3] != 0) {
+                if (page_pixels[((((size_t)(y - k) * page_w) + x) * 4) + 3] != 0) {
                     blocked_n = true;
                 }
             }
             for (uint32_t k = 1; y + k < page_h && !blocked_s; k++) {
-                if (page_pixels[((size_t)(y + k) * page_w + x) * 4 + 3] != 0) {
+                if (page_pixels[((((size_t)(y + k) * page_w) + x) * 4) + 3] != 0) {
                     blocked_s = true;
                 }
             }
             for (uint32_t k = 1; x >= k && !blocked_w; k++) {
-                if (page_pixels[((size_t)y * page_w + (x - k)) * 4 + 3] != 0) {
+                if (page_pixels[((((size_t)y * page_w) + (x - k)) * 4) + 3] != 0) {
                     blocked_w = true;
                 }
             }
             for (uint32_t k = 1; x + k < page_w && !blocked_e; k++) {
-                if (page_pixels[((size_t)y * page_w + (x + k)) * 4 + 3] != 0) {
+                if (page_pixels[((((size_t)y * page_w) + (x + k)) * 4) + 3] != 0) {
                     blocked_e = true;
                 }
             }
@@ -4037,6 +4037,7 @@ void test_atlas_shape_concave_rejects_extrude(void) {
     EXPECT_BUILD_ASSERT(ctx, nt_builder_begin_atlas(ctx, "poly", &opts));
 }
 
+// NOLINTNEXTLINE(readability-function-cognitive-complexity)
 void test_atlas_round_trip_basic(void) {
     (void)MKDIR(TMP_DIR);
     NtBuilderContext *ctx = nt_builder_start_pack(TMP_DIR "/atlas_rt_basic.ntpack");
@@ -4220,8 +4221,8 @@ void test_atlas_round_trip_vertices(void) {
     const uint8_t *ptr = ablob + sizeof(NtAtlasHeader) + ((size_t)ahdr->page_count * sizeof(uint64_t));
     const NtAtlasRegion *regions = (const NtAtlasRegion *)ptr;
     for (uint32_t r = 0; r < ahdr->region_count; r++) {
-        uint32_t vertex_end = (uint32_t)regions[r].vertex_start + regions[r].vertex_count;
-        uint32_t index_end = (uint32_t)regions[r].index_start + regions[r].index_count;
+        uint32_t vertex_end = regions[r].vertex_start + regions[r].vertex_count;
+        uint32_t index_end = regions[r].index_start + regions[r].index_count;
         TEST_ASSERT_TRUE(vertex_end <= ahdr->total_vertex_count);
         TEST_ASSERT_TRUE(index_end <= ahdr->total_index_count);
         TEST_ASSERT_EQUAL_UINT32(0, (uint32_t)regions[r].index_count % 3U);
